# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Repair Shop Database"
  description: "Mock mechanic shop db"
  version: "1.2.0"
schemes:
- "http"
host: "127.0.0.1:5000"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /inventory/:
    post:
      tags:
        - Inventory
        - Method Protection → Mechanic Token
      summary: "Create inventory item"
      description: "Required fields: product_name, price | optional field: recallable, defaults to true"
      
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/CreateInventoryPayload"

      security:
        - bearerAuth: []

      responses:
        201:
          description: Successful creation
          schema:
            $ref: "#/definitions/CreateInventoryResponse"
        400:
          description: "Bad request"
          schema:
            example:
              {
                "message": "'error details' - all inventory fields required"
              }
        401:
          description: "Non-unique product_name field"
          schema:
            example:
              {
                "error": "Products require unique names for clarity."
              }
    get:
      tags:
          - Inventory
          - Method Protection → Mechanic Token
      summary: "Get All Inventory - PROTECTED METHOD - add 'Bearer ' and paste in the mechanic token"
      description: "Mechanics/Admin only\n *Example* /customers/?page=1&per_page=5"

      parameters:
        - in: query
          name: page
          type: integer
          required: false
        - in: query
          name: per_page
          type: integer
          required: false
      
      security:
        - bearerAuth: []

      responses:
        200:
          description: "All customers view for mechanics, pagination is optional, "
          schema:
            $ref: "#/definitions/ViewInventoriesAsMechanic"
  /inventory/{inventory_id}:
    get:
      tags:
        - Inventory
        - Method Protection → Mechanic Token
      summary: "Get inventory item by id number - PROTECTED METHOD - add 'Bearer ' and paste in the mechanic token"
      description: "Mechanic route to get inventory item by id"

      security:
        - bearerAuth: []

      parameters:
        - in: path
          name: inventory_id
          type: integer
          required: true
          description: "The inventory ID number"

      responses:
        200:
          description: "Inventory data shown in a successful response"
          schema:
            $ref: "#/definitions/CreateInventoryResponse"
            example:
              application/json:
                {
                  "id": 1,
                  "product_name": "Regular Oil - HOUSE",
                  "price": 13.45,
                  "recallable": false,
                  "recalled": false,
                  "no_longer_used": false
                }
        404:
          description: "Inventory not found"
          schema:
            example:
              {
                "error": "Inventory not found."
              }
    put:
      tags:
        - Inventory
        - Method Protection → Mechanic Token
      summary: "Get inventory item by id number - PROTECTED METHOD - add 'Bearer ' and paste in the mechanic token"
      description: "Mechanic route to get inventory item by id"

      security:
        - bearerAuth: []

      parameters:
        - in: path
          name: inventory_id
          type: integer
          required: true
          description: "The inventory ID number"

      responses:
        200:
          description: "Inventory data shown in a successful response"
          schema:
            $ref: "#/definitions/CreateInventoryResponse"
            example:
              application/json:
                {
                  "id": 1,
                  "product_name": "Regular Oil - HOUSE",
                  "price": 13.45,
                  "recallable": false,
                  "recalled": false,
                  "no_longer_used": false
                }
        400:  
          description: "Bad request"
          schema:
            example:
              {
                "message": "'error details' - all inventory fields required"
              }
        404:
          description: "Inventory not found"
          schema:
            example:
              {
                "error": "Inventory not found."
              }
    delete:
      tags:
        - Inventory
        - Method Protection → Mechanic Token
      summary: "Delete inventory item by id number - PROTECTED METHOD - add 'Bearer ' and paste in the mechanic token"
      description: "Mechanic route to delete inventory item by id"

      security:
        - bearerAuth: []

      parameters:
        - in: path
          name: inventory_id
          type: integer
          required: true
          description: "The inventory ID number"

      responses:
        200:
          description: "Inventory item successfully deleted"
          schema:
            example:
              {
                "message": "Inventory item successfully deleted"
              }
        404:
          description: "Inventory not found"
          schema:
            example:
              {
                "error": "Inventory not found."
              }
  /inventory/current:
    get:
      tags:
        - Inventory
        - Method Protection → Mechanic Token
      summary: "Get current inventory items"
      description: "Mechanics/Admin only"

      security:
        - bearerAuth: []

      responses:
        200:
          description: "Current inventory items"
          schema:
            $ref: "#/definitions/ViewInventoriesAsMechanic"
            example:
              application/json:
                [
                  {
                    "id": 1,
                    "product_name": "Regular Oil - HOUSE",
                    "price": 13.45,
                    "recallable": false,
                    "recalled": false,
                    "no_longer_used": false
                  },
                  {
                    "id": 2,
                    "product_name": "High Mileage Oil - HOUSE",
                    "price": 15.95,
                    "recallable": false,
                    "recalled": false,
                    "no_longer_used": false
                  }
                ]
  /inventory/search:
    get:
      tags:
        - Inventory
        - Method Protection → Mechanic Token
      summary: "Search inventory items by name"
      description: "Mechanics/Admin only"

      security:
        - bearerAuth: []

      parameters:
        - in: query
          name: product_name
          type: string
          required: false
        - in: query
          name: recallable
          type: integer
          enum: [0, 1]
          required: false
        - in: query
          type: integer
          enum: [0, 1]
          required: false
        - in: query
          name: na
          type: integer
          enum: [0, 1]
          required: false
        - in: query
          name: any
          type: string
          required: false

      responses:
        200:
          description: "Search results for inventory items"
          schema:
            $ref: "#/definitions/ViewInventoriesAsMechanic"
            example:
              application/json:
                [
                  {
                    "id": 1,
                    "product_name": "Regular Oil - HOUSE",
                    "price": 13.45,
                    "recallable": false,
                    "recalled": false,
                    "no_longer_used": false
                  }
                ]
        400:
          description: "Bad request - no search parameters provided"
          schema:
            example:
              {
                "message": "No search parameters provided."
              }
...
  /inventory/shop/product/{product_id}:
    get:
      tags:
        - Inventory
      summary: "Get inventory item by id number for shop"
      description: "Customers can view a specific inventory item by id number."

      parameters:
        - in: path
          name: product_id
          type: integer
          required: true
          description: "The inventory ID number"

      responses:
        200:
          description: "Inventory item data shown in a successful response"
          schema:
            $ref: "#/definitions/CreateInventoryResponse"
            example:
              application/json:
                {
                  "product_name": "Regular Oil - HOUSE",
                  "price": 13.45
                }
        404:
          description: "Inventory item not found"
          schema:
            example:
              {
                "error": "Inventory item not found."
              }
  /inventory/shop/search:
    get:
      tags:
        - Inventory
      summary: "Search inventory items for shop"
      description: "Customers can search inventory items available in the shop.\n*current inventory only*"

      parameters:
        - in: query
          name: name
          type: string
          required: false
        - in: query
          name: price
          type: integer
          required: false
        - in: query
          name: any
          type: string
          required: false

      responses:
        200:
          description: "Search results for inventory items"
          schema:
            $ref: "#/definitions/ViewInventoriesAsCustomer"
        400:
          description: "Bad request - no search parameters provided"
          schema:
            example:
              {
                "message": "No search parameters provided."
              }

# ======================= MECHANICS ====================
...
definitions: 
...
  CreateInventoryPayload:
    type: object
    properties:
      product_name:
        type: string
      price:
        type: number
        format: float
      recallable:
        type: boolean
    example:
      product_name: Regular Oil - HOUSE
      price: 13.45
      recallable: false
...
  CreateInventoryResponse:
    type: object
    properties: 
      id:
        type: integer
      product_name:
        type: string
      price:
        type: number
        format: float
      recalled:
        type: boolean
      recallable:
        type: boolean
      # soft deleted?
      no_longer_used:
        type: boolean
    example:
      id: 1
      product_name: Regular Oil - HOUSE
      price: 13.45
      recallable: false
      recalled: false
      no_longer_used: 0
...
  ViewInventoryAsCustomer:
    type: object
    description: "Inventory item view for customers"
    properties:
      product_name:
        type: string
      price:
        type: number
        format: float
    example:
      {
        "product_name": "Regular Oil - HOUSE",
        "price": 13.45
      }
  ViewInventoriesAsCustomer:
    type: array
    description: "view inventory items product_name and price"
    items:
      $ref: "#/definitions/ViewInventoryAsCustomer"
    example:
      [
        {
          "product_name": "Regular Oil - HOUSE",
          "price": 13.45
        },
        {
          "product_name": "High Mileage Oil - HOUSE",
          "price": 15.95
        },
        {
          "product_name": "Regular Oil - Brand: x",
          "price": 18.0
        },
        {
          "product_name": "High Milage Oil - HOUSE",
          "price": 20.99
        },
        {
          "product_name": "Tires - Brand: Firestone",
          "price": 127.85
        }
      ]
  ViewInventoriesAsMechanic:
    type: array
    description: "pagination optional via ?page=*integer*&per_page=*integer*"
    items:
      $ref: "#/definitions/CreateInventoryResponse"
    example:
      [
        {
          "id": 1,
          "product_name": "Regular Oil - HOUSE",
          "price": 13.45,
          "recallable": false,
          "recalled": false,
          "no_longer_used": 0
        },
        {
          "id": 2,
          "product_name": "High Mileage Oil - HOUSE",
          "price": 15.95,
          "recallable": false,
          "recalled": false,
          "no_longer_used": 0
        },
        {
          "id": 3,
          "product_name": "Regular Oil - Brand: x",
          "price": 18.0,
          "recallable": false,
          "recalled": false,
          "no_longer_used": 0
        },
        {
          "id": 4,
          "product_name": "High Milage Oil - HOUSE",
          "price": 20.99,
          "recallable": false,
          "recalled": false,
          "no_longer_used": false
        },
        {
          "id": 5,
          "product_name": "Tires - Brand: Firestone",
          "price": 127.85,
          "recallable": true,
          "recalled": false,
          "no_longer_used": false
        }
      ]
...
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header